Remove hardcoded path to "gen/gl_API.dtd" subdir.
One of the XML files refers to gl_API.dtd via the relative path
"../gen/gl_API.dtd".  This worked when the XML files were in Mesa (and
they were in a folder called "gen"), but it won't work if we want to
be able to share the XML files between projects.